+++ app/Listeners/Contracts/DeliveryHandler.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: antarus66
+ * Date: 8/20/15
+ * Time: 10:08 AM
+ */
+
+namespace App\Listeners\Contracts;
+
+abstract class DeliveryHandler {
+    protected $delivery;
+}
