+++ app/Listeners/Contracts/WebSocketDeliveryHandler.php
@@ -1,20 +1,14 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: antarus66
- * Date: 8/20/15
- * Time: 10:08 AM
- */
 
-namespace App\Listeners;
+namespace App\Listeners\Contracts;
 
 use App\WebSocket\Contracts\AbstractWebSocketFactory;
 
 /*
  *  Http-server -> WAMP-server delivery
  */
-abstract class DeliveryHandler {
-    protected $delivery;
+class WebSocketDeliveryHandler extends DeliveryHandler
+{
     /**
      * Create the event listener.
      *
