+++ app/Listeners/DeliveryHandlers/WebSocketHandlers/AnswerToBroadcasting.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Listeners;
+namespace App\Listeners\DeliveryHandlers\WebSocketHandlers;
 
 use App\Events\AnswerWasAdded;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use App\Listeners\Contracts\WebSocketDeliveryHandler;
 
-class AnswerToBroadcasting extends DeliveryHandler
+class AnswerToBroadcasting extends WebSocketDeliveryHandler
 {
     /**
      * Handle the event.
