+++ app/Listeners/DeliveryHandlers/WebSocketHandlers/CommentToBroadcasting.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Listeners;
+namespace App\Listeners\DeliveryHandlers\WebSocketHandlers;
 
 use App\Events\CommentWasAdded;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use App\Listeners\Contracts\WebSocketDeliveryHandler;
 
-class CommentToBroadcasting extends DeliveryHandler
+class CommentToBroadcasting extends WebSocketDeliveryHandler
 {
     /**
      * Handle the event.
