+++ app/Listeners/DeliveryHandlers/WebSocketHandlers/FolderRemovingToBroadcasting.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Listeners;
+namespace App\Listeners\DeliveryHandlers\WebSocketHandlers;
 
 use App\Events\FolderWasRemoved;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use App\Listeners\Contracts\WebSocketDeliveryHandler;
 
-class FolderRemovingToBroadcasting extends DeliveryHandler
+class FolderRemovingToBroadcasting extends WebSocketDeliveryHandler
 {
     /**
      * Handle the event.
