+++ app/Listeners/DeliveryHandlers/WebSocketHandlers/FolderUpdatingToBroadcasting.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Listeners;
+namespace App\Listeners\DeliveryHandlers\WebSocketHandlers;
 
 use App\Events\FolderWasUpdated;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use App\Listeners\Contracts\WebSocketDeliveryHandler;
 
-class FolderUpdatingToBroadcasting extends DeliveryHandler
+class FolderUpdatingToBroadcasting extends WebSocketDeliveryHandler
 {
     /**
      * Handle the event.
