+++ app/Listeners/DeliveryHandlers/WebSocketHandlers/QuestionToBroadcasting.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Listeners;
+namespace App\Listeners\DeliveryHandlers\WebSocketHandlers;
 
 use App\Events\QuestionWasAdded;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use App\Listeners\Contracts\WebSocketDeliveryHandler;
 
-class QuestionToBroadcasting extends DeliveryHandler
+class QuestionToBroadcasting extends WebSocketDeliveryHandler
 {
     /**
      * Handle the event.
