+++ app/Listeners/DeliveryHandlers/WebSocketHandlers/VoteRemovingToBroadcasting.php
@@ -1,31 +1,32 @@
 <?php
 
-namespace App\Listeners;
+namespace App\Listeners\DeliveryHandlers\WebSocketHandlers;
 
-use App\Events\VoteWasAdded;
+use App\Events\VoteWasRemoved;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use App\Listeners\Contracts\WebSocketDeliveryHandler;
 
-class VoteToBroadcasting extends DeliveryHandler
+class VoteRemovingToBroadcasting extends WebSocketDeliveryHandler
 {
     /**
      * Handle the event.
      *
-     * @param  VoteWasAdded  $event
+     * @param  VoteWasRemoved  $event
      * @return void
      */
-    public function handle(VoteWasAdded $event)
+    public function handle(VoteWasRemoved $event)
     {
         // to 'entry/{id}/votes' topic
         $this->delivery->send([
-            'data'  => ['post' => $event->vote],
+            'data'  => ['delete' => $event->vote],
             'topic' =>'entry/' . $event->vote->q_and_a_id . '/votes'
         ]);
 
         // to the topic 'entries/{id}' - for model updating
         $this->delivery->send([
             'data'  => [
-                'calls' => ['voteAdd' => $event->vote]
+                'calls' => ['voteDelete' => $event->vote]
             ],
             'topic' =>'entries/' . $event->vote->q_and_a_id
         ]);
