+++ app/Listeners/DeliveryHandlers/WebSocketHandlers/VoteToBroadcasting.php
@@ -1,31 +1,32 @@
 <?php
 
-namespace App\Listeners;
+namespace App\Listeners\DeliveryHandlers\WebSocketHandlers;
 
-use App\Events\VoteWasRemoved;
+use App\Events\VoteWasAdded;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use App\Listeners\Contracts\WebSocketDeliveryHandler;
 
-class VoteRemovingToBroadcasting extends DeliveryHandler
+class VoteToBroadcasting extends WebSocketDeliveryHandler
 {
     /**
      * Handle the event.
      *
-     * @param  VoteWasRemoved  $event
+     * @param  VoteWasAdded  $event
      * @return void
      */
-    public function handle(VoteWasRemoved $event)
+    public function handle(VoteWasAdded $event)
     {
         // to 'entry/{id}/votes' topic
         $this->delivery->send([
-            'data'  => ['delete' => $event->vote],
+            'data'  => ['post' => $event->vote],
             'topic' =>'entry/' . $event->vote->q_and_a_id . '/votes'
         ]);
 
         // to the topic 'entries/{id}' - for model updating
         $this->delivery->send([
             'data'  => [
-                'calls' => ['voteDelete' => $event->vote]
+                'calls' => ['voteAdd' => $event->vote]
             ],
             'topic' =>'entries/' . $event->vote->q_and_a_id
         ]);
